+++ src/fft.cpp
@@ -1,6 +1,6 @@
 #include <fft.h>
 
-torch::Tensor fft(std::vector<double> v1)
+torch::Tensor fft(std::vector<float> v1)
 {
     torch::Tensor input = torch::tensor(v1, torch::kFloat);
     torch::Tensor rfft_result = torch::fft::rfft(input).conj();
@@ -8,9 +8,9 @@ torch::Tensor fft(std::vector<double> v1)
     return rfft_result;
 }
 
-std::vector<double> linspace(double start, double end, size_t num)
+std::vector<float> linspace(float start, float end, size_t num)
 {
-    std::vector<double> result;
+    std::vector<float> result;
     result.reserve(num);
 
     if (num == 0) {return result;}
@@ -20,7 +20,7 @@ std::vector<double> linspace(double start, double end, size_t num)
         return result;
     }
 
-    double step = (end - start) / (num - 1);
+    float step = (end - start) / (num - 1);
     for (size_t i = 0; i < num; ++i) 
     {
         result.push_back(start + i * step);
@@ -29,28 +29,28 @@ std::vector<double> linspace(double start, double end, size_t num)
     return result;
 }
 
-std::vector<double> construct_freqs(std::vector<double> t)
+std::vector<float> construct_freqs(std::vector<float> t)
 {
     int N = t.size();
     int n = N/2 + 1;
 
-    double t_min = *std::min_element(t.begin(), t.end());
-    double t_max = *std::max_element(t.begin(), t.end());
-    double freq_end = N / (t_max - t_min);
+    float t_min = *std::min_element(t.begin(), t.end());
+    float t_max = *std::max_element(t.begin(), t.end());
+    float freq_end = N / (t_max - t_min);
     
 
-    std::vector<double> freqs = linspace(0.0, freq_end, N);
-    std::vector<double> wfreqs(freqs.begin(), freqs.begin() + n);
+    std::vector<float> freqs = linspace(0.0, freq_end, N);
+    std::vector<float> wfreqs(freqs.begin(), freqs.begin() + n);
 
     return wfreqs;
 }
 
-std::vector<double> pos2time(std::vector<double> t)
+std::vector<float> pos2time(std::vector<float> t)
 {
-    std::vector<double> times(t.size());
+    std::vector<float> times(t.size());
     // std::transform(t.begin(), t.end(), times.begin(), [](double x){return x * 2e-3 / C;});  // position here is mm 
 
-    std::transform(t.begin(), t.end(), times.begin(), [](double x){return x * 1e-12;});  // temp use for Kun's data, the time is ps here need to convert to s
+    std::transform(t.begin(), t.end(), times.begin(), [](float x){return x * 1e-12;});  // temp use for Kun's data, the time is ps here need to convert to s
     
     return times;
 }
